Implemented lightgate code, removed temperature sensor code (description)
Currently the temperature sensor appears to create a reading bottleneck in the code, which impedes the accuracy of the RPM that we see because there is a delay in the results being returned.

Therefore I removed commented out the temperature code, which means that collect_data.py does not currently work as intended because there is a one-off index error.